go ahead and move template model to theme dir
